Create Paired_Analysis_Example
Here's an example script for a paired watershed analysis. Several areas for improvement would be to only need to load one .csv with the paired data for all constituents and grab the data needed for each constituent from the one table. Would also be better to declare the name of the control and the treatment watersheds at the top, and then have it automatically grab the right data. It is easy to mix up where to put the control and treatment watersheds, and would make it easier to set up at the top and just run.